requirements.txt
python-telegram-bot~=22.0
pymongo~=4.6.3
colorlog~=6.9.0
dotenv~=0.9.9
python-dotenv~=1.1.0
cryptography~=44.0.2
pandas~=2.2.3
requests~=2.31.0
beautifulsoup4~=4.13.4
Flask~=2.3
xlrd~=2.0.1
openpyxl~=3.1.2